+++ common/src/main/java/dev/gimme/netherreset/Main.java
@@ -1,5 +1,6 @@
 package dev.gimme.netherreset;
 
+import dev.gimme.netherreset.application.EntityHandler;
 import dev.gimme.netherreset.application.PlayerAttachmentAccessor;
 import dev.gimme.netherreset.application.PlayerHandler;
 import dev.gimme.netherreset.domain.inventory.InventoryManager;
@@ -17,10 +18,12 @@ public class Main {
 
     private final ServerConfig serverConfig;
     private final PlayerHandler playerHandler;
+    private final EntityHandler entityHandler;
 
     private Main(PlayerAttachmentAccessor playerAttachmentAccessor) {
         this.serverConfig = new FcapServerConfig();
         this.playerHandler = new PlayerHandler(new InventoryManager(playerAttachmentAccessor, serverConfig));
+        this.entityHandler = new EntityHandler(serverConfig);
     }
 
     public ServerConfig getServerConfig() {
@@ -29,5 +32,9 @@ public class Main {
 
     public PlayerHandler getPlayerHandler() {
         return playerHandler;
+    }
+
+    public EntityHandler getEntityHandler() {
+        return entityHandler;
     }
 }